requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
PySide6==6.6.0
cryptography==41.0.7

# Security and encryption
libsodium==0.7.12
PyNaCl==1.5.0

# Image processing and media scrubbing
opencv-python==4.8.1.78
Pillow==10.1.0
imagehash==4.3.1

# Machine learning and analysis
python-Levenshtein==0.23.0
scikit-learn==1.3.2

# Geolocation and PII detection
geotext==0.4.0
pyap==0.3.2

# Report generation
jinja2==3.1.2
markdown==3.5.1
weasyprint==60.1

# Development and testing
pytest==7.4.3
pytest-qt==4.2.0
ruff==0.1.6
mypy==1.7.1
pre-commit==3.5.0 